requirements.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
Babel==2.9.1
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==1.0.0a5
crashtest==0.3.1
cryptography==37.0.4
distlib==0.3.6
dulwich==0.20.46
filelock==3.8.0
frozenlist==1.3.3
greenlet==2.0.1
html5lib==1.1
idna==3.4
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.16.0
keyring==23.9.3
lockfile==0.12.2
magic-filter==1.0.9
Mako==1.2.4
MarkupSafe==2.1.1
more-itertools==8.14.0
msgpack==1.0.4
multidict==6.0.4
packaging==21.3
pexpect==4.8.0
pipenv==2022.11.30
pkginfo==1.8.3
platformdirs==2.5.2
poetry==1.2.2
poetry-core==1.3.2
poetry-plugin-export==1.2.0
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pycparser==2.21
pylev==1.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.21.0
pytz==2022.7
redis==4.4.0
requests==2.28.1
requests-toolbelt==0.9.1
SecretStorage==3.3.3
shellingham==1.5.0
SQLAlchemy==1.4.45
tomlkit==0.11.5
urllib3==1.26.12
virtualenv==20.17.1
virtualenv-clone==0.5.7
webencodings==0.5.1
yarl==1.8.2
